requirements.txt
autogen
python-dotenv
openai